Feat: 모집글 ssr로 전환, 리팩토링
