+++ src/app/my-page/interests/page.tsx
@@ -1,12 +1,11 @@
 'use client'
 import useAxiosWithAuth from '@/api/config'
-import { ProjectType } from '@/app/panel/MainPage'
 import CuButton from '@/components/CuButton'
 import useInfiniteScroll from '@/hook/useInfiniteScroll'
 import useMedia from '@/hook/useMedia'
 import useModal from '@/hook/useModal'
 import useToast from '@/hook/useToast'
-import { IMainCard } from '@/types/IPostDetail'
+import { IMainCard, ProjectType } from '@/types/IPostDetail'
 import { AlertColor, Stack, Tab, Tabs, Typography } from '@mui/material'
 import React, { useEffect, useState } from 'react'
 import useSWR from 'swr'
